fixes : #1177 added more test cases
