+++ test/apis/utils.go
@@ -10,8 +10,10 @@ import (
 	"strings"
 	"testing"
 
+	framework "github.com/devfile/alizer/pkg/apis/enricher/framework/java"
 	"github.com/devfile/alizer/pkg/apis/model"
 	"github.com/devfile/alizer/pkg/apis/recognizer"
+	"github.com/stretchr/testify/assert"
 )
 
 func updateContent(filePath string, data []byte) error {
@@ -19,7 +21,7 @@ func updateContent(filePath string, data []byte) error {
 	if err != nil {
 		return err
 	}
-	defer func(){
+	defer func() {
 		if err := f.Close(); err != nil {
 			fmt.Printf("error closing file: %s", err)
 		}
@@ -102,4 +104,88 @@ func getTestProjectPath(folder string) string {
 	_, b, _, _ := runtime.Caller(0)
 	basepath := filepath.Dir(b)
 	return filepath.Join(basepath, "..", "..", "resources/projects", folder)
+}
+
+func testOpenLibertyDetector_DoPortsDetection(t *testing.T, projectType string, expectedPorts []int) {
+	tempDir, err := os.MkdirTemp("", projectType+"-test")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.RemoveAll(tempDir)
+
+	hardcodedXML := `
+	<server>
+		<httpEndpoint host="*" httpPort="1234" httpsPort="1235" id="defaultHttpEndpoint"/>
+	</server>
+	`
+	writeTestFile(t, tempDir, "hardcoded_server.xml", hardcodedXML)
+
+	variableXML := `
+	<server>
+		<variable name="default.http.port" defaultValue="9080"/>
+		<variable name="default.https.port" defaultValue="9443"/>
+		<httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="${default.https.port}" id="defaultHttpEndpoint"/>
+	</server>
+	`
+	writeTestFile(t, tempDir, "variable_server.xml", variableXML)
+
+	mixedXML := `
+	<server>
+		<variable name="default.http.port" defaultValue="9080"/>
+		<httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="1235" id="defaultHttpEndpoint"/>
+	</server>
+	`
+	writeTestFile(t, tempDir, "mixed_server.xml", mixedXML)
+
+	emptyVarXML := `
+	<server>
+		<variable name="default.http.port" defaultValue=""/>
+		<variable name="default.https.port" defaultValue="9443"/>
+		<httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="${default.https.port}" id="defaultHttpEndpoint"/>
+	</server>
+	`
+	writeTestFile(t, tempDir, "empty_var_server.xml", emptyVarXML)
+
+	detector := framework.OpenLibertyDetector{}
+	ctx := context.TODO()
+
+	t.Run("Hardcoded Ports", func(t *testing.T) {
+		component := &model.Component{
+			Path: filepath.Join(tempDir, "hardcoded_server.xml"),
+		}
+		detector.DoPortsDetection(component, &ctx)
+		assert.Equal(t, []int{1234, 1235}, component.Ports)
+	})
+
+	t.Run("Variable-based Ports", func(t *testing.T) {
+		component := &model.Component{
+			Path: filepath.Join(tempDir, "variable_server.xml"),
+		}
+		detector.DoPortsDetection(component, &ctx)
+		assert.Equal(t, expectedPorts, component.Ports)
+	})
+
+	t.Run("Mixed Hardcoded and Variable Ports", func(t *testing.T) {
+		component := &model.Component{
+			Path: filepath.Join(tempDir, "mixed_server.xml"),
+		}
+		detector.DoPortsDetection(component, &ctx)
+		assert.Equal(t, []int{9080, 1235}, component.Ports)
+	})
+
+	t.Run("Empty Variable Port", func(t *testing.T) {
+		component := &model.Component{
+			Path: filepath.Join(tempDir, "empty_var_server.xml"),
+		}
+		detector.DoPortsDetection(component, &ctx)
+		assert.Equal(t, []int{9443}, component.Ports)
+	})
+}
+
+func writeTestFile(t *testing.T, dir, filename, content string) {
+	path := filepath.Join(dir, filename)
+	err := os.WriteFile(path, []byte(content), 0644)
+	if err != nil {
+		t.Fatal(err)
+	}
 }
